+++ Util/src/Arr/ArrayLike.scala
@@ -344,7 +344,7 @@ trait ArrayLike[+A] extends Any with ArrayLikeBase[A @uncheckedVariance]
   def toStrsSemiParenth(fToStr: A => String): String = toStrsSemiFold(fToStr).enParenth
 }
 
-case class ArrayLikeShow[A, R <: ArrayLike[A]](evA: Show[A]) extends ShowSeqLike[A, R]
+case class ArrayLikeShow[A, R <: ArrayLike[A]](evA: ShowT[A]) extends ShowSeqLike[A, R]
 { def showComma(obj: R): String = obj.toStrsCommaFold((obj: A) => evA.showT(obj, 0))
   def showSemi(obj: R): String = obj.toStrsSemiFold(evA.showComma(_))
 }
