+++ Util/src/Extensions/ShowTypeExtensions.scala
@@ -1,7 +1,7 @@
 package ostrat
 
 /** The stringer implicit class gives extension methods for Show methods from the implicit Show instance type A. */
-class ShowerTypeExtensions[-A](ev: Show[A], thisVal: A)
+class ShowerTypeExtensions[-A](ev: ShowT[A], thisVal: A)
 { /** Provides the standard string representation for the object. */
   @inline def str: String = ev.showT(thisVal, 10)
