+++ Util/src/Persist/Persist.scala
@@ -5,7 +5,7 @@ import pParse._
 /** The essential persistence type class. it implements both a Show style type class interface, the production of a String representation of the value
   * but also produces an EMon[T] from a String. It Persists and builds objects of type T from CRON syntax. So for example the IntImplicit object in the
   * Persist companion object persists Integers and constructs Integers from Strings. */
-trait Persist[T] extends Show[T] with UnShow[T]
+trait Persist[T] extends ShowT[T] with UnShow[T]
 
 /** Companion object for the persistence type class. The implicit instances for Scala standard library types are to be put in Show. Some methods still
  * in the Persist companion class. */
