+++ Util/src/Persist/PersistSum.scala
@@ -2,10 +2,10 @@
 package ostrat
 import reflect.ClassTag, pParse._
 
-abstract class ShowSum2[ST <: AnyRef, A1 <: ST, A2 <: ST]()(implicit val ct1: ClassTag[A1], val ct2: ClassTag[A2]) extends Show[ST]
+abstract class ShowSum2[ST <: AnyRef, A1 <: ST, A2 <: ST]()(implicit val ct1: ClassTag[A1], val ct2: ClassTag[A2]) extends ShowT[ST]
 {
-  def ev1: Show[A1]
-  def ev2: Show[A2]
+  def ev1: ShowT[A1]
+  def ev2: ShowT[A2]
 
   override def showT(obj: ST, decimalPlaces: Int): String = obj match
   { case a1: A1 => ev1.showT(a1, 0)
@@ -32,11 +32,11 @@ abstract class ShowSum2[ST <: AnyRef, A1 <: ST, A2 <: ST]()(implicit val ct1: Cl
 
 object ShowSum2
 {
-  def apply[ST <: AnyRef, A1 <: ST, A2 <: ST](typeIn: String, ev1In: Show[A1], ev2In: Show[A2])(implicit ct1: ClassTag[A1], ct2: ClassTag[A2]):
+  def apply[ST <: AnyRef, A1 <: ST, A2 <: ST](typeIn: String, ev1In: ShowT[A1], ev2In: ShowT[A2])(implicit ct1: ClassTag[A1], ct2: ClassTag[A2]):
     ShowSum2[ST, A1, A2] = new ShowSum2[ST, A1, A2]()(ct1, ct2)
   { override def typeStr: String = typeIn
-    override def ev1: Show[A1] = ev1In
-    override def ev2: Show[A2] = ev2In
+    override def ev1: ShowT[A1] = ev1In
+    override def ev2: ShowT[A2] = ev2In
   }
 }
