+++ Util/src/Persist/ShowCase.scala
@@ -3,17 +3,17 @@ package ostrat
 
 /** The base trait for the persistence of Case classes, aka Product types */
 trait ShowCase[R] extends ShowCompound[R]
-{ def showMems(): Arr[Show[_]]
+{ def showMems(): Arr[ShowT[_]]
   final override def syntaxDepth: Int = showMems().fMax(_.syntaxDepth) + 1
   def showSemiNames(obj: R): String
   def showCommaNames(obj: R): String
 }
 
 /** Show type class for 1 parameter case classes. */
-abstract class Show1[A1, R](val typeStr: String, name1: String, fArg1: R => A1, val opt1: Option[A1] = None)(implicit ev1: Show[A1], eq1: Eq[A1])
+abstract class Show1[A1, R](val typeStr: String, name1: String, fArg1: R => A1, val opt1: Option[A1] = None)(implicit ev1: ShowT[A1], eq1: Eq[A1])
   extends EqCase1[A1, R](fArg1) with ShowCase[R]
 {
-  final override def showMems(): Arr[Show[_]] = Arr(ev1)
+  final override def showMems(): Arr[ShowT[_]] = Arr(ev1)
   def showSemi(obj: R): String = ev1.showComma(fArg1(obj))
   def showComma(obj: R): String = ev1.showT(fArg1(obj), 10)
   override def showSemiNames(obj: R): String = name1 -:- ev1.showComma(fArg1(obj))
@@ -22,11 +22,11 @@ abstract class Show1[A1, R](val typeStr: String, name1: String, fArg1: R => A1, 
 
 /** Show type class for 2 parameter case classes. */
 class Show2[A1, A2, R](val typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, val opt2: Option[A2] = None,
-  opt1In: Option[A1] = None)(implicit ev1: Show[A1], ev2: Show[A2], eq1: Eq[A1], eq2: Eq[A2]) extends EqCase2[A1, A2, R](fArg1, fArg2) with
+  opt1In: Option[A1] = None)(implicit ev1: ShowT[A1], ev2: ShowT[A2], eq1: Eq[A1], eq2: Eq[A2]) extends EqCase2[A1, A2, R](fArg1, fArg2) with
   ShowCase[R]
 {
   val opt1: Option[A1] = ife(opt2.nonEmpty, opt1In, None)
-  final override def showMems(): Arr[Show[_]] = Arr(ev1, ev2)
+  final override def showMems(): Arr[ShowT[_]] = Arr(ev1, ev2)
   override def showSemi(obj: R): String = ev1.showComma(fArg1(obj)) + "; " + ev2.showComma(fArg2(obj))
   override def showComma(obj: R): String = ev1.showT(fArg1(obj), 10) + ", " + ev2.showT(fArg2(obj), 10)
   override def showSemiNames(obj: R): String = name1 -:- ev1.showComma(fArg1(obj)) + "; " + name2 -:- ev2.showComma(fArg2(obj))
@@ -35,13 +35,13 @@ class Show2[A1, A2, R](val typeStr: String, name1: String, fArg1: R => A1, name2
 
 /** Show type class for 3 parameter case classes. */
 class Show3[A1, A2, A3, R](val typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
-  val opt3: Option[A3] = None, opt2In: Option[A2] = None, opt1In: Option[A1] = None)(implicit ev1: Show[A1], ev2: Show[A2], ev3: Show[A3],
-  eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3]) extends EqCase3[A1, A2, A3, R](fArg1, fArg2, fArg3) with ShowCase[R]
+  val opt3: Option[A3] = None, opt2In: Option[A2] = None, opt1In: Option[A1] = None)(implicit ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3],
+                                                                                     eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3]) extends EqCase3[A1, A2, A3, R](fArg1, fArg2, fArg3) with ShowCase[R]
 {
   val opt2: Option[A2] = ife(opt3.nonEmpty, opt2In, None)
   val opt1: Option[A1] = ife(opt2.nonEmpty, opt1In, None)
   val defaultNum = ife3(opt3.isEmpty, 0, opt2.isEmpty, 1, opt1.isEmpty, 2, 3)
-  override def showMems(): Arr[Show[_]] = Arr(ev1, ev2, ev3)
+  override def showMems(): Arr[ShowT[_]] = Arr(ev1, ev2, ev3)
 
   final override def showSemi(obj: R): String =
   { val p1 = fArg1(obj)
@@ -78,14 +78,14 @@ class Show3[A1, A2, A3, R](val typeStr: String, name1: String, fArg1: R => A1, n
 /** Show type class for 4 parameter case classes. */
 abstract class Show4[A1, A2, A3, A4, R](val typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String,
   fArg3: R => A3, name4: String, fArg4: R => A4, val opt4: Option[A4] = None, opt3In: Option[A3] = None, opt2In: Option[A2] = None,
-  opt1In: Option[A1] = None)(implicit ev1: Show[A1], ev2: Show[A2],  ev3: Show[A3], ev4: Show[A4], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3],
-  eq4: Eq[A4]) extends EqCase4[A1, A2, A3, A4, R](fArg1, fArg2, fArg3, fArg4)(eq1, eq2, eq3, eq4) with ShowCase[R]
+  opt1In: Option[A1] = None)(implicit ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3], ev4: ShowT[A4], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3],
+                             eq4: Eq[A4]) extends EqCase4[A1, A2, A3, A4, R](fArg1, fArg2, fArg3, fArg4)(eq1, eq2, eq3, eq4) with ShowCase[R]
 {
   val opt3: Option[A3] = ife(opt4.nonEmpty, opt3In, None)
   val opt2: Option[A2] = ife(opt3.nonEmpty, opt2In, None)
   val opt1: Option[A1] = ife(opt2.nonEmpty, opt1In, None)
 
-  final override def showMems(): Arr[Show[_]] = Arr(ev1, ev2, ev3, ev4)
+  final override def showMems(): Arr[ShowT[_]] = Arr(ev1, ev2, ev3, ev4)
 
   override def showSemi(obj: R): String =
   { val p1 = fArg1(obj)
@@ -113,8 +113,8 @@ abstract class Show4[A1, A2, A3, A4, R](val typeStr: String, name1: String, fArg
 /** Show type class for 5 parameter case classes. */
 class Show5[A1, A2, A3, A4, A5, R](val typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
   name4: String, fArg4: R => A4, name5: String, fArg5: R => A5, val opt5: Option[A5], opt4In: Option[A4] = None, opt3In: Option[A3] = None,
-  opt2In: Option[A2] = None, opt1In: Option[A1] = None)(implicit ev1: Show[A1], ev2: Show[A2], ev3: Show[A3], ev4: Show[A4], ev5: Show[A5],
-  eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5]) extends EqCase5[A1, A2, A3, A4, A5, R](fArg1, fArg2, fArg3, fArg4, fArg5) with
+  opt2In: Option[A2] = None, opt1In: Option[A1] = None)(implicit ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3], ev4: ShowT[A4], ev5: ShowT[A5],
+                                                        eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5]) extends EqCase5[A1, A2, A3, A4, A5, R](fArg1, fArg2, fArg3, fArg4, fArg5) with
   ShowCase[R]
 {
   val opt4: Option[A4] = ife(opt5.nonEmpty, opt4In, None)
@@ -122,7 +122,7 @@ class Show5[A1, A2, A3, A4, A5, R](val typeStr: String, name1: String, fArg1: R 
   val opt2: Option[A2] = ife(opt3.nonEmpty, opt2In, None)
   val opt1: Option[A1] = ife(opt2.nonEmpty, opt1In, None)
 
-  final override def showMems(): Arr[Show[_]] = Arr(ev1, ev2, ev3, ev4, ev5)
+  final override def showMems(): Arr[ShowT[_]] = Arr(ev1, ev2, ev3, ev4, ev5)
 
   override def showSemi(obj: R): String =
   { val p1 = fArg1(obj)
@@ -153,8 +153,8 @@ object Show5
 {
   def apply[A1, A2, A3, A4, A5, R](typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
     name4: String, fArg4: R => A4, name5: String, fArg5: R => A5, opt5: Option[A5] = None, opt4: Option[A4] = None, opt3: Option[A3] = None,
-    opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: Show[A1], ev2: Show[A2], ev3: Show[A3], ev4: Show[A4], ev5: Show[A5], eq1: Eq[A1],
-    eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5]) =
+    opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3], ev4: ShowT[A4], ev5: ShowT[A5], eq1: Eq[A1],
+                                                      eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5]) =
     new Show5[A1, A2, A3, A4, A5, R](typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, name5, fArg5, opt5, opt4, opt3, opt2, opt1)(
     ev1, ev2, ev3, ev4, ev5, eq1, eq2, eq3, eq4: Eq[A4], eq5)
 }
@@ -163,8 +163,8 @@ object Show5
 class Show6[A1, A2, A3, A4, A5, A6, R](val typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String,
   fArg3: R => A3, name4: String, fArg4: R => A4, name5: String, fArg5: R => A5, name6: String, fArg6: R => A6, val opt6: Option[A6],
   val opt5In: Option[A5] = None, opt4In: Option[A4] = None, opt3In: Option[A3] = None, opt2In: Option[A2] = None, opt1In: Option[A1] = None)(
-  implicit ev1: Show[A1], ev2: Show[A2], ev3: Show[A3], ev4: Show[A4], ev5: Show[A5], ev6: Show[A6], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3],
-  eq4: Eq[A4], eq5: Eq[A5], eq6: Eq[A6]) extends EqCase6[A1, A2, A3, A4, A5, A6, R](fArg1, fArg2, fArg3, fArg4, fArg5, fArg6) with ShowCase[R]
+                                        implicit ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3], ev4: ShowT[A4], ev5: ShowT[A5], ev6: ShowT[A6], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3],
+                                        eq4: Eq[A4], eq5: Eq[A5], eq6: Eq[A6]) extends EqCase6[A1, A2, A3, A4, A5, A6, R](fArg1, fArg2, fArg3, fArg4, fArg5, fArg6) with ShowCase[R]
 {
   val opt5: Option[A5] = ife(opt6.nonEmpty, opt5In, None)
   val opt4: Option[A4] = ife(opt5.nonEmpty, opt4In, None)
@@ -172,7 +172,7 @@ class Show6[A1, A2, A3, A4, A5, A6, R](val typeStr: String, name1: String, fArg1
   val opt2: Option[A2] = ife(opt3.nonEmpty, opt2In, None)
   val opt1: Option[A1] = ife(opt2.nonEmpty, opt1In, None)
 
-  final override def showMems(): Arr[Show[_]] = Arr(ev1, ev2, ev3, ev4, ev5, ev6)
+  final override def showMems(): Arr[ShowT[_]] = Arr(ev1, ev2, ev3, ev4, ev5, ev6)
 
   override def showSemi(obj: R): String =
   { val p1 = fArg1(obj)
@@ -207,8 +207,8 @@ object Show6
   def apply[A1, A2, A3, A4, A5, A6, R](typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
     name4: String, fArg4: R => A4, name5: String, fArg5: R => A5, name6: String, fArg6: R => A6, opt6: Option[A6] = None, opt5: Option[A5] = None,
     opt4: Option[A4] = None, opt3: Option[A3] = None, opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit
-    ev1: Show[A1], ev2: Show[A2], ev3: Show[A3], ev4: Show[A4], ev5: Show[A5], ev6: Show[A6],
-    eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5], eq6: Eq[A6]) =
+                                                                                                        ev1: ShowT[A1], ev2: ShowT[A2], ev3: ShowT[A3], ev4: ShowT[A4], ev5: ShowT[A5], ev6: ShowT[A6],
+                                                                                                        eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5], eq6: Eq[A6]) =
     new Show6[A1, A2, A3, A4, A5, A6, R](typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, name5, fArg5, name6, fArg6,
     opt6, opt5, opt4, opt3, opt2, opt1)(ev1, ev2, ev3, ev4, ev5, ev6, eq1, eq2, eq3, eq4, eq5, eq6)
 }
