+++ Util/src/pParse/Token.scala
@@ -21,7 +21,7 @@ trait Token extends TextSpan
 
 object Token
 {
-  implicit val showImplicit: Show[Token] = new ShowSimple[Token]("Token")
+  implicit val showImplicit: ShowT[Token] = new ShowSimple[Token]("Token")
   { def showT(obj: Token, decimalPlaces: Int): String = obj.str
   }
 }
